requirements/base.txt
Django==4.2.2
django-auditlog==2.3.0
django-captcha==1.0.0
django-celery-beat==2.5.0
django-ckeditor==6.6.1
django-cors-headers==4.1.0
django-filter==23.2
django-import-export==3.2.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-modeltranslation==0.18.10
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-recaptcha==3.0.0
django-redis==5.3.0
django-resized==1.0.2
django-timezone-field==5.1
djangorestframework==3.14.0
drf-yasg==1.21.6
Jinja2==3.1.2
phonenumbers==8.13.15
Pillow==9.5.0
psycopg2-binary==2.9.6
python-crontab==2.7.1
python-dateutil==2.8.2
python-environ==0.4.54
redis==4.6.0
requests==2.31.0
djangorestframework-simplejwt==5.2.2
coreapi==2.3.3
sentry_sdk==1.4.3
unidecode==1.3.2
django-rosetta==0.9.9
dj-rest-auth==2.2.5
social-auth-app-django==5.0.0
social-auth-core==4.3.0
django-allauth==0.51.0
opencv-python==4.8.0.74
numpy==1.24.4
sorl-thumbnail==12.9.0
django_silk==5.0.4
